Cleanup useless stuff
Remove placeholder directories.
Add entries in .gitignore for files created at runtime.

Also: Add help message because I always forget the env
variable to set.

+++ fetch/owncloud/fetch.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "$1" = "-h" ]; then
+	echo "KEY_FETCH=<the key> ./fetch.sh"
+	exit 1
+fi
+
 set -e
 
 [[ -e .tmp ]] && rm -rf .tmp
